fix video enabled in movie and tv show details and add saved,rating dialog

+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/reusableinteractionlistener/addtofavourite/AddToFavouriteInteractionListener.kt
@@ -1,7 +1,7 @@
 package com.berlin.aflami.viewmodel.reusableinteractionlistener.addtofavourite
 
 interface AddToFavouriteInteractionListener {
-    fun onAddMediaToFavouriteListClicked(favouriteListId: Int, mediaId: Int)
+    fun onAddMediaToFavouriteListClicked(favouriteListId: Int, mediaId: Long)
     fun onSelectFavouriteList(favouriteListId: Int)
     fun onCreateNewFavouriteListClicked()
     fun onCancelAddingToFavouriteClicked()
